REFACTOR: component API files renamed
	- Now components API have the same name, inside their respective
	  component folder.
