+++ components/WindowBox/componentAPI.ts
