+++ components/WindowBox/index.tsx
@@ -1,7 +1,7 @@
 import redirectFromWindowToNavItem from "@/accessibilityScripts/redirectFromWindowToNavItem"
 import trapFocus from "@/accessibilityScripts/trapFocus"
 import setWindowChildrenTabIndexNegative from "@/accessibilityScripts/windowChildrenTabIndexHandler"
-import WindowsAPI from "@/components/WindowBox/windowBoxAPI"
+import WindowsAPI from "@/components/WindowBox/componentAPI"
 import WindowContentArea from "@/components/WindowBox/WindowContentArea"
 import WindowLineBreak from "@/components/WindowBox/WindowLineBreak"
 import WindowTaskbar from "@/components/WindowBox/WindowTaskbar"
